Add APIs for setting OAuth and basic auth headers to the ClientBuilder (#2680)
Motivation:
It would be nice if we provide APIs for setting basic auth, OAuth 1.0 and OAuth 2.0 header as we support those in the server-side.

Modifications:
- Add `auth(BasicToken)`, `auth(OAuth1aToken)` and `auth(OAuth2Token)` to `AbstractClientOptionsBuilder`.
- Fix a bug where OAuth1a paremters are not decoded when they are percent-encoded.

Result:
- You can now easily set the authorization header.
- (Breaking) `BasicToken`, `OAuth1aToken` and `OAuth2Token` classes are moved to `com.linecorp.armeria.common.auth`.
