+++ core/src/main/java/com/linecorp/armeria/server/auth/AuthTokenExtractors.java
@@ -20,6 +20,9 @@ import java.util.function.Function;
 
 import com.linecorp.armeria.common.HttpHeaderNames;
 import com.linecorp.armeria.common.RequestHeaders;
+import com.linecorp.armeria.common.auth.BasicToken;
+import com.linecorp.armeria.common.auth.OAuth1aToken;
+import com.linecorp.armeria.common.auth.OAuth2Token;
 
 /**
  * A utility class that provides singleton instances of authorization token extractor functions.
