+++ core/src/main/java/com/linecorp/armeria/server/auth/OAuth2TokenExtractor.java
@@ -31,6 +31,7 @@ import com.google.common.base.Strings;
 
 import com.linecorp.armeria.common.HttpHeaderNames;
 import com.linecorp.armeria.common.RequestHeaders;
+import com.linecorp.armeria.common.auth.OAuth2Token;
 
 import io.netty.util.AsciiString;
