+++ core/src/test/java/com/linecorp/armeria/common/auth/OAuth1aTokenTest.java
@@ -13,14 +13,14 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.server.auth;
+package com.linecorp.armeria.common.auth;
 
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_CONSUMER_KEY;
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_NONCE;
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_SIGNATURE;
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_SIGNATURE_METHOD;
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_TIMESTAMP;
-import static com.linecorp.armeria.server.auth.OAuth1aToken.OAUTH_TOKEN;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_CONSUMER_KEY;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_NONCE;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_SIGNATURE;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_SIGNATURE_METHOD;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_TIMESTAMP;
+import static com.linecorp.armeria.common.auth.OAuth1aToken.OAUTH_TOKEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
@@ -49,31 +49,31 @@ class OAuth1aTokenTest {
                                                             .put("x-others", "g")
                                                             .build()));
         assertThat(token).isNotEqualTo(OAuth1aToken.of(ImmutableMap.<String, String>builder()
-                                                            .put(OAUTH_CONSUMER_KEY, "1")
-                                                            .put(OAUTH_TOKEN, "2")
-                                                            .put(OAUTH_SIGNATURE_METHOD, "3")
-                                                            .put(OAUTH_SIGNATURE, "4")
-                                                            .put(OAUTH_TIMESTAMP, "5")
-                                                            .put(OAUTH_NONCE, "6")
-                                                            .put("x-others", "7")
-                                                            .build()));
+                                                               .put(OAUTH_CONSUMER_KEY, "1")
+                                                               .put(OAUTH_TOKEN, "2")
+                                                               .put(OAUTH_SIGNATURE_METHOD, "3")
+                                                               .put(OAUTH_SIGNATURE, "4")
+                                                               .put(OAUTH_TIMESTAMP, "5")
+                                                               .put(OAUTH_NONCE, "6")
+                                                               .put("x-others", "7")
+                                                               .build()));
         assertThat(token).isNotEqualTo(OAuth1aToken.of(ImmutableMap.<String, String>builder()
-                                                            .put(OAUTH_CONSUMER_KEY, "a")
-                                                            .put(OAUTH_TOKEN, "b")
-                                                            .put(OAUTH_SIGNATURE_METHOD, "c")
-                                                            .put(OAUTH_SIGNATURE, "d")
-                                                            .put(OAUTH_TIMESTAMP, "0")
-                                                            .put(OAUTH_NONCE, "f")
-                                                            .build()));
+                                                               .put(OAUTH_CONSUMER_KEY, "a")
+                                                               .put(OAUTH_TOKEN, "b")
+                                                               .put(OAUTH_SIGNATURE_METHOD, "c")
+                                                               .put(OAUTH_SIGNATURE, "d")
+                                                               .put(OAUTH_TIMESTAMP, "0")
+                                                               .put(OAUTH_NONCE, "f")
+                                                               .build()));
         assertThat(token).isNotEqualTo(OAuth1aToken.of(ImmutableMap.<String, String>builder()
-                                                            .put(OAUTH_CONSUMER_KEY, "a")
-                                                            .put(OAUTH_TOKEN, "b")
-                                                            .put(OAUTH_SIGNATURE_METHOD, "c")
-                                                            .put(OAUTH_SIGNATURE, "d")
-                                                            .put(OAUTH_TIMESTAMP, "0")
-                                                            .put(OAUTH_NONCE, "f")
-                                                            .put("x-others", "g")
-                                                            .put("x-others-2", "h")
-                                                            .build()));
+                                                               .put(OAUTH_CONSUMER_KEY, "a")
+                                                               .put(OAUTH_TOKEN, "b")
+                                                               .put(OAUTH_SIGNATURE_METHOD, "c")
+                                                               .put(OAUTH_SIGNATURE, "d")
+                                                               .put(OAUTH_TIMESTAMP, "0")
+                                                               .put(OAUTH_NONCE, "f")
+                                                               .put("x-others", "g")
+                                                               .put("x-others-2", "h")
+                                                               .build()));
     }
 }
