+++ core/src/test/java/com/linecorp/armeria/server/auth/BasicTokenTest.java
@@ -19,6 +19,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 
+import com.linecorp.armeria.common.auth.BasicToken;
+
 class BasicTokenTest {
     @Test
     void testEquals() {
